% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ppi.R
\name{calculate_ppi}
\alias{calculate_ppi}
\title{Function to calculate Poverty Probability Index (PPI) for Kenya}
\usage{
calculate_ppi(
  df,
  category = c("nl100", "nlFood", "nl150", "nl200", "ppp100", "ppp190", "ppp320",
    "ppp550", "ppp800", "ppp125", "ppp250", "ppp500", "percentile20", "percentile40",
    "percentile60", "percentile80")
)
}
\arguments{
\item{df}{Dataframe containing PPI indicator variables for each household.}

\item{category}{Poverty classifications.}
}
\value{
A dataframe of PPI for each household in \code{df}.
}
\description{
Function to calculate Poverty Probability Index (PPI) for Kenya
}
\examples{
#calculate_ppi(df = household)

}
