% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-calculate_ppi_civ.R
\name{get_data_civ}
\alias{get_data_civ}
\title{Get Ivory Coast PPI data from IPA's data analysis and collection tool}
\usage{
get_data_civ(path)
}
\arguments{
\item{path}{Directory path to IPA's Ivory Coast PPI data analysis and
collection XLSX file}
}
\value{
A dataframe of household data based on IPA's Ivory Coast PPI data
  analysis and collection tool
}
\description{
Get Ivory Coast PPI data from IPA's data analysis and collection tool
}
\examples{
\dontrun{
  get_data_col(path = paste(system.file("ppi", package = "ppicalc"), "ivoryCoast.xlsx", sep = "/"))
}

}
