% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-calculate_ppi_col.R
\name{get_data_col}
\alias{get_data_col}
\title{Get Colombia PPI data from IPA's data analysis and collection tool}
\usage{
get_data_col(path)
}
\arguments{
\item{path}{Directory path to IPA's Colombia PPI data analysis and
collection XLSX file}
}
\value{
A dataframe of household data based on IPA's Colombia PPI data
  analysis and collection tool
}
\description{
Get Colombia PPI data from IPA's data analysis and collection tool
}
