% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ppi.R
\name{get_poverty_class}
\alias{get_poverty_class}
\title{Get poverty classification labels given a selected country to obtain PPI for}
\usage{
get_poverty_class(ccode)
}
\arguments{
\item{ccode}{A character value for three letter ISO code for country to
calculate PPI for. Currently, 12 countries are supported by this function:
1) Colombia (COL); 2) Ivory Coast (CIV); 3) Dominican Republic (DOM);
4) Ghana (GHA); 5) Kenya (KEN); 6) Mozambique (MOZ); 7) Myanmar (MMR);
8) Peru (PER); 9) Philippines (PHL); 10) Rwanda (RWA); 11) Senegal (SEN);
12) Togo (TGO)}
}
\value{
A vector of names/labels of poverty classifications
}
\description{
Get poverty classification labels given a selected country to obtain PPI for
}
\examples{

get_poverty_class(ccode = "COL")

}
