% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi_cohort.R
\name{get_ppi_cohort}
\alias{get_ppi_cohort}
\title{get_ppi_cohort}
\usage{
get_ppi_cohort(data = NULL, index, ccode)
}
\arguments{
\item{data}{A data frame that contains data on PPI scores.}

\item{index}{The households' PPI score of class numeric (default). If \code{data}
parameter is provided, then \code{index} should be a character value
indicating the variable name in \code{data} containing the households'
PPI score.}

\item{ccode}{Three letter ISO country code}
}
\value{
A data frame of households' poverty probabilities based on various poverty
    metrics
}
\description{
Get a cohort or sample of households' poverty probability based on each
household's Poverty Probability Index (PPI) score. This function uses PPI
lookup tables for 60 countries available via the \code{ppitables} package.
}
\examples{
# Apply function on a dataset from Bangladesh that includes PPI scores from
# the washdata package
\dontrun{
get_ppi_cohort(data = indicatorsDataBGD,
               index = "ppi",
               ccode = "BGD")
}

# Apply function on a dataset from Bangladesh but specifying vector for PPI
# and keeping data as NULL
\dontrun{
get_ppi_cohort(index = indicatorsDataBGD$ppi,
               ccode = "BGD")
}

}
