% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi_quintile.R
\name{get_ppi_quintile}
\alias{get_ppi_quintile}
\title{get_ppi_quintile}
\usage{
get_ppi_quintile(ppi, probs = c(0.2, 0.4, 0.6, 0.8, 1))
}
\arguments{
\item{ppi}{A numeric vector of poverty probabilities or likelihoods}

\item{probs}{A numeric vector between 0 to 1 of quantile probabilities.
Default values are a vector of equal probabilities (0.2, 0.4, 0.6, 0.8, 1).}
}
\value{
A numeric vector between 1 to 5 classifying data into quintiles of
    increasing poverty probabilities.
}
\description{
Group population by quintiles based on their poverty probability
}
\examples{
#

}
