% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-calculate_ppi_civ.R
\name{ppi_civ_ipa}
\alias{ppi_civ_ipa}
\title{Function to calculate Poverty Probability Index (PPI) for Ivory Coast using
IPA's new PPI methodology}
\usage{
ppi_civ_ipa(
  path,
  category = names(ppitables::ppiCIV2018)[2:ncol(ppitables::ppiCIV2018)]
)
}
\arguments{
\item{path}{Directory path to IPA's Ivory Coast PPI data analysis and
collection XLSX file}

\item{category}{Poverty classifications that can be calculated for
Ivory Coast using IPA's new PPI methodology. By default, all 15 poverty
classifications for Ivory Coast are specified}
}
\value{
A dataframe of PPI for each household in \code{df}.
}
\description{
Function to calculate Poverty Probability Index (PPI) for Ivory Coast using
IPA's new PPI methodology
}
\examples{
\dontrun{
  ppi_col_ipa(path = paste(system.file("ppi", package = "ppicalc"),
                           "ivoryCoast.xlsx", sep = "/"))
}

}
