% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-calculate_ppi_col.R
\name{ppi_col_ipa}
\alias{ppi_col_ipa}
\title{Function to calculate Poverty Probability Index (PPI) for Colombia using
IPA's new PPI methodology}
\usage{
ppi_col_ipa(
  path,
  category = names(ppitables::ppiCOL2018)[2:ncol(ppitables::ppiCOL2018)]
)
}
\arguments{
\item{path}{Directory path to IPA's Colombia PPI data analysis and
collection XLSX file}

\item{category}{Poverty classifications that can be calculated for
Colombia using IPA's new PPI methodology. By default, all 18 poverty
classifications for Colombia are specified (nl100)}
}
\value{
A dataframe of PPI for each household in \code{df}.
}
\description{
Function to calculate Poverty Probability Index (PPI) for Colombia using
IPA's new PPI methodology
}
\examples{
\dontrun{
  ppi_col_ipa(df = get_data_col(path = "colombia.xlsx"))
}

}
