% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintile_ppi.R
\name{quintile_ppi}
\alias{quintile_ppi}
\title{Group population by quintiles based on their poverty probability}
\usage{
quintile_ppi(ppi, probs = c(0.2, 0.4, 0.6, 0.8, 1))
}
\arguments{
\item{ppi}{A numeric vector of poverty probabilities or likelihoods.}

\item{probs}{A numeric vector between 0 to 1 of quantile probabilities.
Default values are a vector of equal probabilities (0.2, 0.4, 0.6, 0.8, 1).}
}
\value{
A numeric vector between 1 to 5 classifying data into quintiles of
    increasing poverty probabilities.
}
\description{
Group population by quintiles based on their poverty probability
}
\examples{
# ppi based on USAID extreme poverty definition using Bangladesh data from
# washdata package and PPI lookup table from ppitables package
# install.packages("washdata"); library(washdata)
ppi <- get_ppi_cohort(index = washdata::indicatorsDataBGD$ppi,
                      ppiTable = ppitables::ppiBGD2013)
quintile_ppi(ppi = ppi$extreme)

}
