% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quintile_ppi_cohort.R
\name{quintile_ppi_cohort}
\alias{quintile_ppi_cohort}
\title{Group population by quintiles based on their multiple poverty probabilities}
\usage{
quintile_ppi_cohort(ppi, probs = c(0.2, 0.4, 0.6, 0.8, 1))
}
\arguments{
\item{ppi}{A data frame of poverty probabilities or likelihoods.}

\item{probs}{A numeric vector between 0 to 1 of quantile probabilities.
Default values are a vector of equal probabilities (0.2, 0.4, 0.6, 0.8, 1).}
}
\value{
A data frame of multiple quintile classifications of between 1 to 5
    of increasing poverty probabilities.
}
\description{
Group population by quintiles based on their multiple poverty probabilities
}
\examples{
# ppi for all poverty definitions using Bangladesh data from
# washdata package and PPI lookup table from ppitables package
# install.packages("washdata"); library(washdata)
ppi <- get_ppi_cohort(index = washdata::indicatorsDataBGD$ppi,
                      ppiTable = ppitables::ppiBGD2013)
quintile_ppi_cohort(ppi = ppi)

}
