% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-score_ppi_cohort.R
\name{score_ppi_cohort}
\alias{score_ppi_cohort}
\title{Get a cohort or sample of households' poverty probability score based on
their responses to the ten country-specific questions.}
\usage{
score_ppi_cohort(data, ccode)
}
\arguments{
\item{data}{A data frame containing responses to the 10 questions used to
elicit information for calculating the PPI score}

\item{ccode}{Three letter ISO code for a country}
}
\value{
A numeric vector of PPI scores for each household
}
\description{
Get a cohort or sample of households' poverty probability score based on
their responses to the ten country-specific questions.
}
\examples{
#
\dontrun{
score_ppi_cohort(data = surveyDataBGD, ccode = "BGD")
}

}
