% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ppi.R
\name{calculate_ppi}
\alias{calculate_ppi}
\title{Function to calculate Poverty Probability Index (PPI) for Kenya}
\usage{
calculate_ppi(path, ccode, category = get_poverty_class(ccode = ccode))
}
\arguments{
\item{path}{Directory path to IPA's country-specific PPI data analysis and
collection XLSX file}

\item{ccode}{A character value for three letter ISO code for country to
calculate PPI for. Currently, 12 countries are supported by this function:
1) Colombia (COL); 2) Ivory Coast (CIV); 3) Dominican Republic (DOM);
4) Ghana (GHA); 5) Kenya (KEN); 6) Mozambique (MOZ); 7) Myanmar (MMR);
8) Peru (PER); 9) Philippines (PHL); 10) Rwanda (RWA); 11) Senegal (SEN);
12) Togo (TGO)}

\item{category}{Poverty classifications. This defaults to all the poverty
classifications available for the selected country}
}
\value{
A dataframe of PPI for each household in \code{df}.
}
\description{
Function to calculate Poverty Probability Index (PPI) for Kenya
}
\examples{
calculate_ppi(path = paste(system.file("ppi", package = "ppicalc"),
                           "colombia.xlsx", sep = "/"),
              ccode = "COL")

}
