% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi.R
\name{get_ppi}
\alias{get_ppi}
\title{get_ppi}
\usage{
get_ppi(index, ccode)
}
\arguments{
\item{index}{The household's PPI score}

\item{ccode}{Three letter ISO country code}
}
\value{
A vector of household's poverty probabilities based on various poverty
    metrics
}
\description{
Get a household's poverty probability based on their Poverty Probability
Index (PPI) score. This function uses PPI lookup tables for 60 countries
available via the \code{ppitables} package.
}
\examples{
# Apply function on a household in Bangladesh (BGD) with a PPI score of 65
\dontrun{
get_ppi(index = 65, ccode = "BGD")
}

}
