% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ppi_cohort.R
\name{get_ppi_cohort}
\alias{get_ppi_cohort}
\title{get_ppi_cohort}
\usage{
get_ppi_cohort(index, ppiTable)
}
\arguments{
\item{index}{The households' PPI score of class numeric.}

\item{ppiTable}{Name of country PPI table to use from the \code{ppitables}
package}
}
\value{
A data frame of households' poverty probabilities based on various poverty
    metrics
}
\description{
Get a cohort or sample of households' poverty probability based on each
household's Poverty Probability Index (PPI) score. This function uses PPI
lookup tables for 60 countries available via the \code{ppitables} package.
}
\examples{
# Apply function on a dataset from Bangladesh that includes PPI scores from
# the washdata package
\dontrun{
get_ppi_cohort(data = washdata::indicatorsDataBGD,
               index = "ppi",
               ppiTable = ppiBGD2013)
}

# Apply function on a dataset from Bangladesh but specifying vector for PPI
# and keeping data as NULL
#\dontrun{
get_ppi_cohort(index = washdata::indicatorsDataBGD$ppi,
               ppiTable = ppitables::ppiBGD2013)
#}

}
